Moving the logging and log class to public api
