+++ driver/src/main/java/org/neo4j/driver/internal/connector/socket/LoggingResponseHandler.java
@@ -21,7 +21,7 @@ package org.neo4j.driver.internal.connector.socket;
 import java.util.Arrays;
 import java.util.Map;
 
-import org.neo4j.driver.internal.spi.Logger;
+import org.neo4j.driver.v1.Logger;
 import org.neo4j.driver.v1.Value;
 
 import static org.neo4j.driver.internal.messaging.DiscardAllMessage.DISCARD_ALL;
