+++ driver/src/main/java/org/neo4j/driver/internal/logging/JULogging.java
@@ -20,8 +20,8 @@ package org.neo4j.driver.internal.logging;
 
 import java.util.logging.Level;
 
-import org.neo4j.driver.internal.spi.Logger;
-import org.neo4j.driver.internal.spi.Logging;
+import org.neo4j.driver.v1.Logger;
+import org.neo4j.driver.v1.Logging;
 
 public class JULogging implements Logging
 {
