+++ driver/src/main/java/org/neo4j/driver/v1/Config.java
@@ -22,7 +22,6 @@ import java.io.File;
 import java.util.logging.Level;
 
 import org.neo4j.driver.internal.logging.JULogging;
-import org.neo4j.driver.internal.spi.Logging;
 import org.neo4j.driver.v1.util.Immutable;
 
 import static java.lang.System.getProperty;
