+++ driver/src/main/java/org/neo4j/driver/v1/Logger.java
@@ -16,9 +16,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.driver.internal.spi;
+package org.neo4j.driver.v1;
 
-public interface Logging
+public interface Logger
 {
-    Logger getLog( String name );
+    void error( String message, Throwable cause );
+
+    void info( String message, Object... params );
+
+    void warn( String message, Object... params );
+
+    void debug( String message, Object... params );
+
+    void trace( String message, Object... params );
+
+    boolean isTraceEnabled();
+
+    boolean isDebugEnabled();
 }
